[stablehlo] Implement product of householder reflectors (#15555)
The stablehlo.custom call for product of householder reflectors can be
directly generated via multiple operations. Implemented the non-blocked
variant as a lowering directly to `linalg`. Testing we have a simplified
lit test as the algorithm is not friendly for lit tests. Instead we use a
numerical validation in stablehlo_ops e2e tests.