[VectorDistribution] Add patterns for distributing transfer_read/transfer_write (#16115)
This patch adds distribution patterns for transfer_read/transfer_write
and lowers them to vector.load/vector.store per thread.

These distribution patterns do the lowering for
transfer_read/transfer_write in one-shot, which is different from how
transfer_read/transfer_write are lowered in upstream mlir. The upstream
patterns unroll one dimension at a time and apply the patterns
recursively. We do this lowering in one-shot because we have the layout
attribute which defines the iteration space for the lowering.