NFC: [Codegen][GPU] Split GPUTensorAlloc into tiling and promotion (#16214)
Currently GPUTensorAlloc tiles reduction dimensions as well as doing
promotion, but there is no mechanical reason these two transformations
have to happen at the same time. Additionally, the reduction tiling has
no relevance to the transpose pipeline that uses this pass today.

This split is in preparation for another form of promotion pass that
runs later on.