Switching code to use FunctionOpInterface instead of FuncOp. There's lots of tech debt around func.func, it turns out. As many uses of func (besides creation) have been switched to the FunctionOpInterface as a first step toward support non-func dialects. These changes will also allow more passes to correctly operate on initializers and future dispatch export ops that carry more information than func.func allows. Codegen is welcome to keep using func where it wants as it's more aligned with upstream than the rest of the compiler but there was a lot of unnecessary coupling to func that predated the existence of the FunctionOpInterface/CallOpInterface this cleans up.
There's a few places that still rely on func that don't really need
it but they aren't trivial to switch and the hope is that the bulk
of the cleanup here (passes, basic walks/iteration, etc) will make
them easier to burn down now that they are the exception rather than
the rule.

Progress on #16187.