[Codegen] Add a configuration attribute dict to translation info (#16224)
The attribute as it is today is rather unwieldy, especially when trying
to add new fields to translation info. In general, translation info
needs to provide all of the configuration details for the particular
pass pipeline it specifies, so this moves the software pipelining
configuration attributes to a configuration dictionary so that the
codegen dialect doesn't have to care about the details of a particular
pass pipeline.