Nothing is unreachable (#16261)
This comes from debugging the mysterious crashes on CI on @mariecwhite
's https://github.com/openxla/iree/pull/16259.

Debugged:
https://github.com/openxla/iree/pull/16259#issuecomment-1917417773

The motivation for these unreachable statements was a performance
concern that is no longer current on LLVMCPU now that inlining works
reliably. The switch variables here are in practice always compile-time
constants so the whole switch statement vanishes in post-linking IR
optimization passes. The only places where the unreachable still helps
are: VMVX and ukernel own micro benchmarks and tests, and the impact is
small, so it's not worth keeping.

As a bonus, this makes the fallback now automatic when adding a new
mmt4d type not handled by existing architecture-sepcific code, so this
just makes ukernels easier to extend, as evidenced by the fact that
@mariecwhite did run into this.

One place had to retain an unreachable to work around a specific
compiler bug. As it's specific to one compiler, it doesn't need a
portable macro to wrap around the unreachable builtin.