[CPU][DataTiling] Extend encodings to linalg.generic ops (#16163)
This PR extends data tiling to linalg.generic ops that are contractions.
This allows materializeEncoding to generate the correct pack ops for any
form of generalized batch_matmul operation, avoiding generating extra
transpose ops to be handled by Flow. In order to really avoid this,
`LiftGenericToTransposeBatchMatmulPass` is also removed in this PR.

When `LiftGenericToTransposeBatchMatmulPass` is removed, we need to
handle unsigned semantics for some mmt4d ukernels, so this PR moves the
logic for creating a separate producer op that arith.extui matmul inputs
directly to materializeEncoding.