Drop build-dir outputs from build_all CI job. (#16333)
Progress on https://github.com/openxla/iree/issues/16203

All remaining uses of the `build-dir` outputs have been removed, so we
can skip generating and uploading those files. The `install-dir` is half
the size, and we may be able to shrink it even further (I think we still
want debug symbols and assertions for jobs that test using packages, but
that would be the largest lever to flip).