[Codegen][ROCM] Add a codegen pipeline for supported MFMA variants (#16258)
The codegen pipeline added here, named LLVMGPUVectorDistribute, uses
the vector distribution patterns to do distribution to MFMA types. This
pipeline is still largely work in progress and thus is hidden behind
a flag: --iree-codegen-llvmgpu-use-vector-distribution

However this sets the skeleton for the pipeline. The major additions
here are:

1. The IREEGPUDialect which adds attributes for MFMA types and adds some
   convenience templating/interfaces for defining other kinds of mma
   types (i.e. wmma/mma.sync). This should help normalize many of the
   strategies around these tensor core operations without needing to
   update the code in too many places. Details are often not that nice
   though :)

2. The LLVMGPUVectorDistribute pass pipeline for matmul codegen.
   Currently because vector distribution is a bit tuned for matmul
   cases, this special cases to matmuls, but much of the code still in
   the process of landing should allow for this to be a much more
   general pipeline in the future. For now taking baby steps.

To supplement the pass pipeline, adds GPUVectorAlloc and
LLVMGPUVectorDistribute passes. The vector allocation pass is
somewhat ad-hoc and could be dropped depending on some decisions made
around pipelining, but is needed for now to get promotion. It's
a simple enough pattern to slot in and out depending on how promotion
happens. The vector distribution pass builds off of the recently added
work on that front and implements the logic for setting the anchors for
distribution. That pass is subject to change as layout attributes
evolve.

Things that do not work well in this first iteration:
1. No pipelining
2. Barrier placement is not smart
3. Distribution is limited to a single warp/subgroup
4. Kernel configuration is very basic, however without multi-warp
   distribution it is probably not worth building out the logic there
   just yet.