Add a number of runtime python bindings and refine the HalFence.wait() behavior. (#16371)
* Adds static methods to `HalElementType`: `is_byte_aligned`,
`dense_byte_count`
* Binds the ref protocol for `HalBuffer` so it can be be put in lists.
* Adds `HalBufferView.byte_length`
* Adds `HalFence.signal` and `HalFence.fail`
* Changes behavior of `HalFence.wait` to return False if not satisfied
due to timeout and True if satisfied. Raises a Python exception on
either synchronous or asynchronous failure. The prior behavior raised
exceptions for timeout and reported success on asynchronous failure. The
latter is a bug, and the former is just a bad practice (i.e. don't raise
exceptions for normal behavior).