[rocdl] Create basic code structure for ROCDL CodeGen (#16342)
This commit starts the rework of LLVMGPU compiler backend for future
production. Specifically it starts a set of new source code files for
ROCDL CodeGen configuration. The goal is to start with a minimal working
structure and gradually port over relevant pass pipelines with proper
target handling and heuristics. So in this commit we only pick the warp
reduction pipeline which sort of matches the criteria here. Also we
don't shuffle the passes around yet--the new ROCDL library just depends
on the existing LLVMGPU library and its passes directly to get started.
This is meant to avoid churns and we can do the mechanical change once
everything is sorted.

Progress towards https://github.com/openxla/iree/issues/16341