Only run benchmark workflows when requested. (#16439)
Progress on https://github.com/openxla/iree/issues/16431.

The `build_e2e_test_artifacts` job produces artifacts used for
benchmarking but it has been also providing test coverage that the
programs in the benchmark suite can be compiled correctly. That test
coverage would be a better fit for a test job like `test_gpu` (ci.yml)
or `regression_test_cpu` (pkgci.yml). As it is now, that job is building
multiple very similar configurations and it then uploads those files to
the cloud (to the tune of 70-130GB per CI run).

This also adds back coverage that was missing from
`test_benchmark_suites`. Those tests should be migrated to another
workflow not connected to benchmarks so they can be ran and debugged in
isolation. This setup is too brittle and complex.