[rocdl] Use a full slice of vector values to avoid anchors (#16534)
Fused consumer operations are expected to take the layout of a producer
contraction for the current version of the matmul pipeline. Previously
it was only taking a forward slice of the program to check whether a
tranfer needs a particular layout. This now takes a slice through all
vector values to infer whether a particular transfer will take the
layout of the contraction. This is still only an approximation of what
the layout analysis will do, and is spooky action at a distance, but
this is really the only way to set anchors for the analysis.