[torch] Implement async and mutability programming model. (#16486)
This patch removes the reference torch-mlir
`FuncBackendTypeConversionPass` and implements a new
`FuncConversionPass` which directly targets IREE functions. This
centralizes in one place the conversions that let us get to a high
fidelity between torch->iree, whereas they were spread across half a
dozen places. Centralizing lets us implement both mutability and the
corresponding impact that has to synchronization policy at the ABI. This
relies on knowledge of actual Torch semantics and is unsuitable for the
generic input conversion pipelines.

Also move the conversion from tm_tensor to linalg_ext up to where we
convert into tm_tensor vs being after all conversions are done. The
forcing function was that such op level conversions should not be done
once we commit the ABI.

A key difference vs the previous state: At the end of `--torch-to-iree`,
all functions are defined on BufferView (vs tensor) and have appropriate
import/export/synchronization in place. This was being done in different
stages before. Attributes have been added to converted functions to
disable those default bufferization/wrapping stages when coming in from
Torch in this way, since it has already been fully converted to IREE
semantics.

The effect to all of this is that if you compile a Torch program that
performs an inplace update of a tensor argument or a containing
nn.Module buffer, then the FX importer will note the arrangement, and
IREE will properly set up those arguments to be mutable and tie them so
that they can operate in place on device (subject to certain fusion and
other heuristics which can force a copy). Only tensor content
replacement is supported right now. This may be expanded as Torch moves
forward on these points.