[CPU] Add a specialized pipeline for LinalgExt::AttentionOp. (#16577)
The revision adds a new pipeline for LinalgExt ops. It is an
experimental pipeline, and should eventually get merged into
MultiTilingPipeline.

The new pipeline introduces vector level of tiling to LinalgExt, and
vectorization. Some dimension of attention op is not able to tile at
this moment, so we set all the tile sizes to 1 which avoids huge
vectors. Because the reduction dimension of matmuls is not tiled. Here
is selected IR dump:
https://gist.githubusercontent.com/hanhanW/db4511da681d4932cb81dd68cc98976f/raw/08c3cc42c9d7fb86b769f60dc712fecb9fb10700/dump.mlir

Towards https://github.com/openxla/iree/issues/16421