[hip] Optionally enable graph command buffer and tests (#16604)
This commit made `hipDrvGraphAddMemcpyNode` as an optional
symbol so that even when we are missing it we can still build the hip
HAL driver without problem. Instead it will error out at runtime.
This allows us to start testing the hip graph based command buffer.
For that, tests are registered.