[Flow] Avoid fusion of dequantization-like ops with producers (#16610)
Dequantization-like operations (e.g. arith.extf/extui/extsi) are
generally best to aggressively fuse with consumers because
this minimizes the amount of memory on the boundaries of dispatches.

This patch changes the heuristics in FusionOfTensorOps to avoid fusion
of a dequant-like consumer with its producer. Additionally it disables
multi-use fusion of a multi-use dequant-like producer for the same
reason, namely we would rather fuse by cloning the dequant-like op
instead of typical multi-use fusion.