Delete experimental/cpu_ukernel (#16540)
It gets in the way of the calling convention change that comes next.

It also outlived its usefulness, which was:
- A stepping stone along the route from the initial state where ukernels
where VMVX-only, to the current state where everything just works in
LLVMCPU with ukernels-as-bitcode. Circa May 2023 there was that
intermediate state where ukernels calls worked but only this kind of
plugin could provide the ukernel definitions to link to.
- An extension point used by some external ukernel authors (targeting
other CPU architectures) at a time when bitcode was not as convenient to
supply from the outside. That should be fixed now. Reach out if you need
to get up to speed with how to bring-your-own-bitcode.