Refining emplace allocations to handle dynamic shapes. (#16656)
#15261 was aggressive in disabling the pass for any offset/size SSA
values computed as part of updates. In most programs the offset arith
ops would end up after the dispatches we would want to emplace and this
pass wouldn't touch them. Now we try to move the arith ops up before the
dispatch so we can do the placement.

Progress on #16647. This allows us to emplace llama2 variable updates
with dynamic offsets.