[ROCm] Set option to preload kernel arguments (#16659)
This needs to be enabled both on the compiler and the runtime side; this
PR adds the latter.

Unfortunately, this does not seem to be exposed through the public API
and requires an environment variable to be set before initializing the
hip runtime. This makes this a temporary hack until we can upgrade to
ROCm 6.1 where kernarg is enabled by default.

This brings a few % latency improvement on SDXL's unet.