Making the ElideAsyncCopiesPass support stream.async.slice. (#16667)
This is only looking for some very specific patterns that we can
(mostly) verify are safe. This can definitely be improved and widened to
more cases but this is better than the nothing we do today for slice
ops. It does have some gotchas around potentially increasing lifetime of
source tensors but we don't have a good way of measuring that today so
🤷.

Fixes #16647.