Drop lists of VulkanSPIRV flags. (#16680)
Previously this could take a list of `--iree-vulkan-target-triple` and
`--iree-vulkan-target-env` flags. Other backends only support a single
flag and as we move towards multi-device, we will be standardizing on
device spec files with `#hal.device.target`s and their executable
formats, _not_ loose flags for each backend.

Now only one of those flags can be set at a time and both are a single
value, not a list.

Progress on https://github.com/openxla/iree/issues/15468 (the flag
parsing code in here was not compatible with how compiler plugins are
currently configured).