[LinalgExt] Add scaling to attention op (#16679)
This commit add support for a scale input to the attention
op. This is to follow semantics of scaled dot product
semantics for attention like explained in
https://paperswithcode.com/method/scaled

Right now we only thread through the lower part of the
stack, meaning linalg ext ops and its code generation.
We also need to eventually fix upper layers to get the
scale from the input. For now we only support the PyTorch
default scale case--rsqrt(d).

---------

Co-authored-by: Lei Zhang <antiagainst@gmail.com>