Retry failed pytest cases to try limiting flakes. (#16718)
This uses https://pypi.org/project/pytest-retry/ to retry failed tests.
When we have 1000+ test cases running on real hardware with GPUs /
accelerators, some runs are likely to have flakes. Hopefully failures
like
https://github.com/openxla/iree/actions/runs/8206978856/job/22447547947?pr=16706#step:9:54
will be quieter this way.

~~Unfortunately, that doesn't seem to compose well with
https://pypi.org/project/pytest-timeout/ (a test that times out will not
be retried, from what I can tell).~~ (maybe this isn't the case... can't
quite tell from a few local and CI runs)

Bumped the commit ref to
https://github.com/nod-ai/SHARK-TestSuite/commit/46ab63b271fddd78e071c0590e9a073f6282b005
to include the dep in requirements.txt and a small patch to be more
compatible with pytest-retry.