Add optional attribute to set MFMA read layout (#16733)
This PR adds an optional attribute to set_contraction_layouts that sets
another anchor to the defining op of the LHS and/or RHS operands of the
contraction. The layout anchor is a modified version of the MFMA layout
that reads double the number of elements as the MFMA layout but halves
the batch. The user can specify whether to set it on the LHS(0), RHS(1)
or both.