[Flow] Fix exponential blowup when optimizing dynamic `tensor.dim`s (#16847)
The `simplifyDimOps` function used `reifyDynamicResultDims` to replace
instances of `tensor.dim` in the hopes of it producing more optimized
expressions than a plain `tensor.dim`. In the case that
`reifyDynamicResultDims` failed to do so, it fell back to also
generating `tensor.dim`.

This is problematic for that specific caller as it generates these for
every dynamic dimension. In a test-case of mine this leads to generating
two `tensor.dim` operations for every initial `tensor.dim` which was
then performed for every `linalg.generic` operation, producing roughly
2^20 `tensor.dim` operations.
This could also be seen when running the testcases in
`form_dispatch_regions.mlir` where the output contained exponential as
many unused `tensor.dim` operations as the input.

This PR fixes that by issue by separating the concept of `reifying`
dynamic dimensions and `optimizing` dynamic dimension. The former is
used when an expression representing the dynamic dimension is required
and retains the current behavior of falling back to generating
`tensor.dim` operations. The latter is now used to optimize existing
`tensor.dim`, returning failure if unable to do so.