Ukernels: replace boilerplate by tables. (#16879)
This -650 lines shrink comes from introducing tables (the .inl files)
where we declare the mmt4d tile functions. From that, we can generate
not only the declarations but also the whole `select_tile_func` logic.

While this PR stands on its own, there is a subtext here that in a
subsequent PR (#16880) I'm going to argue --- against all common sense!
--- for actually exporting these tables, as an interim solution to the
problem discussed in #16848 until the better solutions outlined there
can land. Don't want this to be a distraction on this PR as the tables
are not being exported here, and they are justified by independent
reasons here.

That is entirely within the `arch/` subdir. The only other part of this
PR outside of `arch` is a renaming: `mmt4d_tile.c` only contains the
generic (fallback) tile functions, so made its name reflect that.

I tried hard to also drop the whole CMake configure-checks and
configured-headers and for `IREE_UK_BUILD...`, tried moving that back
into the C files as checks of compiler versions, but ran again into the
mystery about the CI jobs `build_test_all_windows` and
`build_test_runtime_windows` having a very different MSVC behavior there
despite being the exact same version, as noted earlier in
https://github.com/openxla/iree/pull/16637.